create branch: hooks
The "hooks" branch is a branch started from a "detached" root - its
history is separate from the main branch for the repository.

This way hook scripts used for this repository can ebe versioned and
"carried" with the repository, but the commits dealing with the hooks
are not part of the git suite's source code.

To install the hooks that exist in this branch, run the "install_hooks"
script ("install-hooks" commit to be added soon).  The script will backup
any existing hook script installed before replacing it.