install-hooks: a script to install git hook scripts into .git/hooks
This script will copy any files that exist in the same directory as the
script tha have the following characteristics:

  - they aren't this script (ie., don't have the same name as this script)
  - they have no extenstion
  - they are executable
  - they are not "hidden" (begin with a ".")

The hook scripts will be copied to .git/hooks/ so that git will pick them up
according to the git-defined filename.

If the script finds that a hook script with the same name already exists in
.git/hooks, the install script will backup the existing file appending
the data and ".bak" to the filename.  In other words, if this script would
overwrite an existing git hook named "pre-commit" then that existing hook
sscript would be renamed to something like "pre-commit.2025-07-01.bak"
depending on the date.

NOTE: this means that running this script several times on the same day
will result in backup getting over written.  Please keep this in mind!